Rename: 폴더 구조 변경 - Event를 사용하는 폴더 구조를 뺐습니다.
