+++ modules/Event/EventHandler.hpp
@@ -1,7 +1,7 @@
 #ifndef EVENTHANDLER_HPP
 # define EVENTHANDLER_HPP
 # include "Kqueue.hpp"
-# include "../UData/HttpResponse/HttpResponseHandler.hpp"
+# include "HttpResponseHandler.hpp"
 class EventHandler
 {
 public:
