+++ modules/webserv/ServManager.hpp
@@ -11,10 +11,10 @@
 #include <cstdio>
 #include <map>
 #include <string>
-#include "Kqueue.hpp"
-#include "../config/ConfParser.hpp"
-# include "../UData/HttpResponse/HttpResponseHandler.hpp"
-# include "EventHandler.hpp"
+# include "../Event/Kqueue.hpp"
+# include "../config/ConfParser.hpp"
+# include "../Event/HttpResponseHandler.hpp"
+# include "../Event/EventHandler.hpp"
 /**
  * @brief kqueue using echo server
  * @details <kqueue 동작 로직>
